TestBed/requirements.txt
pillow==
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
transformers